# Custom marshaler for Proposal