Before
```
//go test github.com/mysteriumnetwork/discovery/proposal/v3 -bench=. -run BenchmarkProposal_UnmarshalBinary -count 10
//goos: darwin
//goarch: amd64
//pkg: github.com/mysteriumnetwork/discovery/proposal/v3
//cpu: Intel(R) Core(TM) i7-4770HQ CPU @ 2.20GHz
//BenchmarkProposal_UnmarshalBinary-8   	      12	  98144248 ns/op	24918711 B/op	       8 allocs/op
//BenchmarkProposal_UnmarshalBinary-8   	      12	  95996542 ns/op	24918710 B/op	       7 allocs/op
//BenchmarkProposal_UnmarshalBinary-8   	      12	  94519130 ns/op	19325925 B/op	       4 allocs/op
//BenchmarkProposal_UnmarshalBinary-8   	      12	  95355681 ns/op	24918693 B/op	       7 allocs/op
//BenchmarkProposal_UnmarshalBinary-8   	      12	  95350196 ns/op	24918740 B/op	       8 allocs/op
//BenchmarkProposal_UnmarshalBinary-8   	      12	  98040917 ns/op	24918686 B/op	       7 allocs/op
//BenchmarkProposal_UnmarshalBinary-8   	      12	  95320773 ns/op	24918732 B/op	       8 allocs/op
//BenchmarkProposal_UnmarshalBinary-8   	      12	  96656234 ns/op	24918685 B/op	       7 allocs/op
//BenchmarkProposal_UnmarshalBinary-8   	      12	  95178269 ns/op	19325925 B/op	       4 allocs/op
//BenchmarkProposal_UnmarshalBinary-8   	      12	  95749498 ns/op	24918567 B/op	       8 allocs/op
//PASS
//ok  	github.com/mysteriumnetwork/discovery/proposal/v3	19.150s
```

After:
```
go test github.com/mysteriumnetwork/discovery/proposal/v3 -bench=. -run BenchmarkProposal_UnmarshalBinary -count 10
goos: darwin
goarch: amd64
pkg: github.com/mysteriumnetwork/discovery/proposal/v3
cpu: Intel(R) Core(TM) i7-4770HQ CPU @ 2.20GHz
BenchmarkProposal_UnmarshalBinary-8   	      24	  48522681 ns/op	103143616 B/op	  330002 allocs/op
BenchmarkProposal_UnmarshalBinary-8   	      24	  48175973 ns/op	103143625 B/op	  330002 allocs/op
BenchmarkProposal_UnmarshalBinary-8   	      21	  49644298 ns/op	103143613 B/op	  330002 allocs/op
BenchmarkProposal_UnmarshalBinary-8   	      24	  51579927 ns/op	103143618 B/op	  330002 allocs/op
BenchmarkProposal_UnmarshalBinary-8   	      21	  50148173 ns/op	103143627 B/op	  330002 allocs/op
BenchmarkProposal_UnmarshalBinary-8   	      24	  49348670 ns/op	103143593 B/op	  330002 allocs/op
BenchmarkProposal_UnmarshalBinary-8   	      24	  49478306 ns/op	103143621 B/op	  330002 allocs/op
BenchmarkProposal_UnmarshalBinary-8   	      22	  47261933 ns/op	103143618 B/op	  330002 allocs/op
BenchmarkProposal_UnmarshalBinary-8   	      25	  48275154 ns/op	103143624 B/op	  330002 allocs/op
BenchmarkProposal_UnmarshalBinary-8   	      25	  46217577 ns/op	103143642 B/op	  330002 allocs/op
PASS
ok  	github.com/mysteriumnetwork/discovery/proposal/v3	16.961s
```